+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+  * Invalid ranges like `1..<0` now produce a run-time error instead
+    of an empty array.
+
 ### Fixed
 
   * Work around (probable) AMD OpenCL compiler bug for